requirements.txt
fastapi
uvicorn
pydantic
python-multipart

torch
torchvision
transformers
sentence-transformers

opencv-python
numpy
faiss-cpu

python-dotenv
Pillow


